Fixing receiver pump open/close race issue and adding a new unit test to cover the fix.
