+++ test/Microsoft.Azure.EventHubs.Tests/Processor/ProcessorTestBase.cs
@@ -202,10 +202,11 @@ namespace Microsoft.Azure.EventHubs.Tests.Processor
                 await Task.Delay(TimeSpan.FromSeconds(60));
 
                 TestUtility.Log("Sending an event to each partition");
+                var ehClient = EventHubClient.CreateFromConnectionString(TestUtility.EventHubsConnectionString);
                 var sendTasks = new List<Task>();
                 foreach (var partitionId in PartitionIds)
                 {
-                    sendTasks.Add(this.SendToPartitionAsync(partitionId, $"{partitionId} event.", TestUtility.EventHubsConnectionString));
+                    sendTasks.Add(TestUtility.SendToPartitionAsync(ehClient, partitionId, $"{partitionId} event."));
                 }
                 await Task.WhenAll(sendTasks);
 
@@ -404,6 +405,8 @@ namespace Microsoft.Azure.EventHubs.Tests.Processor
         {
             var customConsumerGroupName = "cgroup1";
 
+            var ehClient = EventHubClient.CreateFromConnectionString(TestUtility.EventHubsConnectionString);
+
             // Generate a new lease container name that will be used through out the test.
             string leaseContainerName = Guid.NewGuid().ToString();
 
@@ -422,7 +425,6 @@ namespace Microsoft.Azure.EventHubs.Tests.Processor
             {
                 // Create a receiver on the consumer group and try to receive.
                 // Receive call will fail if consumer group is missing.
-                var ehClient = EventHubClient.CreateFromConnectionString(TestUtility.EventHubsConnectionString);
                 var receiver = ehClient.CreateReceiver(customConsumerGroupName, this.PartitionIds.First(), PartitionReceiver.StartOfStream);
                 await receiver.ReceiveAsync(1, TimeSpan.FromSeconds(5));
             }
@@ -479,7 +481,7 @@ namespace Microsoft.Azure.EventHubs.Tests.Processor
                 var sendTasks = new List<Task>();
                 foreach (var partitionId in PartitionIds)
                 {
-                    sendTasks.Add(this.SendToPartitionAsync(partitionId, $"{partitionId} event.", TestUtility.EventHubsConnectionString));
+                    sendTasks.Add(TestUtility.SendToPartitionAsync(ehClient, partitionId, $"{partitionId} event."));
                 }
 
                 await Task.WhenAll(sendTasks);
@@ -925,15 +927,12 @@ namespace Microsoft.Azure.EventHubs.Tests.Processor
                 await Task.Delay(5000);
 
                 TestUtility.Log($"Sending {totalNumberOfEventsToSend} event(s) to each partition");
+                var ehClient = EventHubClient.CreateFromConnectionString(TestUtility.EventHubsConnectionString);
                 var sendTasks = new List<Task>();
                 foreach (var partitionId in PartitionIds)
                 {
-                    for (int i = 0; i < totalNumberOfEventsToSend; i++)
-                    {
-                        sendTasks.Add(this.SendToPartitionAsync(partitionId, $"{partitionId} event.", TestUtility.EventHubsConnectionString));
-                    }
+                    sendTasks.Add(TestUtility.SendToPartitionAsync(ehClient, partitionId, $"{partitionId} event.", totalNumberOfEventsToSend));
                 }
-
                 await Task.WhenAll(sendTasks);
 
                 // Wait until all partitions are silent, i.e. no more events to receive.
@@ -959,20 +958,6 @@ namespace Microsoft.Azure.EventHubs.Tests.Processor
             }
 
             return runResult;
-        }
-
-        protected async Task SendToPartitionAsync(string partitionId, string messageBody, string connectionString)
-        {
-            var eventHubClient = EventHubClient.CreateFromConnectionString(connectionString);
-            try
-            {
-                var partitionSender = eventHubClient.CreatePartitionSender(partitionId);
-                await partitionSender.SendAsync(new EventData(Encoding.UTF8.GetBytes(messageBody)));
-            }
-            finally
-            {
-                await eventHubClient.CloseAsync();
-            }
         }
     }
